pyproject.toml
[tool.poetry]
name = "peak-it-2024-backend"
version = "0.1.0"
description = "Backend for PeakIT hackathon project"
authors = ["toastmanager <mrbonkyou@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
alembic = "^1.13.3"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.12"
ruff = "^0.7.0"
sqladmin = "^0.20.0"
itsdangerous = "^2.2.0"
aioboto3 = "^13.2.0"

[tool.ruff]
target-version = "py310"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
